Update RunTime
about the method reduceSpace : The input would be one line of the input String. First, it reduces all consecutive spaces to only one space. Then check if there is still a white space at the first Char to delete it. To check the first char value, we first need to make sure if the length of that line/string is greater than zero. I changed the method name from "splitLine" to reduceSpace to make more sense.